# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/profile/me": {
      "get": {
        "operationId": "getMyProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/v1/profile": {
      "patch": {
        "operationId": "updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update profile",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/v1/profile/complete": {
      "get": {
        "operationId": "isComplete",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Check whether basic profile is complete",
        "tags": [
          "profile"
        ]
      }
    },
...
  },
  "info": {
    "title": "FOODMISSION Data Framework API",
...
    "version": "1.0.0",
    "contact": {
      "name": "FOODMISSION Team",
      "url": "https://github.com/reedu-reengineering-education/foodmission-data-framework",
      "email": "support@foodmission.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.foodmission.dev",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.foodmission.dev",
      "description": "Staging server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token obtained from Keycloak authentication",
        "in": "header"
      }
    },
    "schemas": {
...
      "ProfileUpdateDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Unique username"
          },
          "age": {
            "type": "number",
            "description": "Age in years"
          },
          "country": {
            "type": "string",
            "description": "Country"
          },
          "region": {
            "type": "string",
            "description": "Region / state / province"
          },
          "zip": {
            "type": "string",
            "description": "ZIP / postal code"
          },
          "language": {
            "type": "string",
            "description": "Language preference"
          },
          "gender": {
            "type": "string",
            "description": "Gender",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "UNSPECIFIED",
              "PREFER_NOT_TO_SAY"
            ]
          },
          "annualIncome": {
            "type": "number",
            "description": "Annual income"
          },
          "educationLevel": {
            "type": "string",
            "description": "Education level"
          },
          "weightKg": {
            "type": "number",
            "description": "Weight in kg"
          },
          "heightCm": {
            "type": "number",
            "description": "Height in cm"
          },
          "activityLevel": {
            "type": "string",
            "description": "Activity level"
          },
          "healthGoals": {
            "type": "object",
            "description": "Health goals (JSON)"
          },
          "nutritionTargets": {
            "type": "object",
            "description": "Nutrition targets (JSON)"
          }
        }
      },
...
    }
  }
}